Added AtMAGNEXMap class
Added the MAGNEX geometry and added a MAGNEX visualizer

Implemented the possibility to specify in which plane the PadPlane is and improved the MAGNEX visualizer.

Removed the pad2channel from AtMAGNEXMap.

Implemented the use of the AtMAGNEXMap class in the AtMAGNEXParsingTask.

Added AtHitClusterEvent class for events that use AtHitClusterFull instead of AtHit.

Added AtMAGNEXParserAndCluster class in order to clusterize MAGNEX hits by row.

Progress in implementing the new AtHitClusterEvent class in the viewer.